+++ src/app/page.tsx
@@ -1,5 +1,5 @@
-import { getProduct } from "@/api/firebase";
-import AddProductBtn from "@/clientComp/AddProductBtn";
+import { getProduct } from "@/services/firebase/product";
+import AddProductBtn from "@/components/AddProductBtn";
 import { Product } from "@/types/products";
 
 export default async function Home() {
